Pull fix for OpenSSL with Perl 5.18+.
OpenSSL v1.0.1e doesn't build with Perl 5.18+ due to some POD errors
in the OpenSSL docs.  I submitted a pull request to fix the problem
(https://github.com/opscode/omnibus-software/pull/57) but it is not
yet merged.  Temporarily peg omnibus-software dependency to my fork
that includes the fix in order to build on latest Arch and other
distros with latest Perl.