Fix vagrant-wrapper after Hashicorp changes
Hashicorp recently deleted their puppet-modules repository which
we were dependent on for the vagrant bin script that correctly
sets up the necessary environment variables for Vagrant and executes
the executable from the gem.  As a quick work-around simply pull a
recent copy of that script, make it available in a new repository
(also called vagrant-wrapper) and switch to using that.

I compared the wrapper script with the current official way of
packaging Vagrant to ensure the script will continue to work with
newer releases.  Whenever the Vagrant team settles on a final
packaging system we can consider replacing this wrapper script
with something based on upstream's work, but in the meantime I
don't anticipate any major difficulties keeping this approach
working.