Update to Vagrant v1.4.3
Upgrade to Vagrant v1.4.3 and add a patch to drop the dependency on
vagrant-spec.  I think vagrant-spec is a good idea and I commend the
Vagrant developers for their efforts on it.  Unfortunately, vagrant-spec
release don't appear to be in sync with vagrant releases; in fact,
vagrant-spec has no releases at all.  Unfortunately, this means that
unless you build your Vagrant packages at the time of the official
release there is no guarantee that the latest changes in vagrant-spec
as of the time you build Vagrant packages will still be compatible
with the Vagrant release you are trying to build.  *sigh*

Just drop the dependency and hope Mitchell, et al get their act
together in the future.